test: keep trace resource lease test scoped
